pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cheetahclaws"
version = "3.05.76"
description = "CheetahClaws: An Extensible, Python-Native Agent System for Autonomous Multi-Model Workflows"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.30.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "pyte>=0.8.0",
]

[project.optional-dependencies]
voice       = ["sounddevice"]
vision      = ["Pillow"]
autosuggest = ["prompt_toolkit>=3.0.43"]
browser     = ["playwright"]
files       = ["pymupdf", "openpyxl"]
ocr         = ["pytesseract", "Pillow"]
trading     = ["yfinance>=0.2.30", "rank-bm25>=0.2.2"]
web         = ["sqlalchemy>=2.0", "bcrypt>=4.0", "PyJWT>=2.8.0"]
all         = ["sounddevice", "Pillow", "prompt_toolkit>=3.0.43", "playwright", "pymupdf", "openpyxl", "pytesseract", "yfinance>=0.2.30", "rank-bm25>=0.2.2", "sqlalchemy>=2.0", "bcrypt>=4.0", "PyJWT>=2.8.0"]

[project.scripts]
cheetahclaws = "cheetahclaws:main"

[tool.setuptools]
py-modules = [
    "cheetahclaws",
    "agent",
    "agent_runner",
    "auxiliary",
    "bootstrap",
    "circuit_breaker",
    "cloudsave",
    "compaction",
    "cc_config",
    "context",
    "error_classifier",
    "id_uniquify",
    "coercion",
    "health",
    "jobs",
    "logging_utils",
    "memory",
    "providers",
    "quota",
    "runtime",
    "session_store",
    "skills",
    "subagent",
    "tmux_tools",
    "tool_registry",
]
packages = [
    "tools",
    "cc_mcp", "memory", "monitor", "multi_agent", "plugin", "skill", "task",
    "voice", "video", "checkpoint",
    "ui", "web", "bridges", "commands",
    "research", "research.sources",
    "modular", "modular.video", "modular.voice", "modular.trading",
    "modular.trading.data", "modular.trading.engines", "modular.trading.agents",
]

[tool.setuptools.package-data]
"web" = ["*.js", "*.css", "*.html", "*.ico", "*.png", "static/js/*.js", "static/*.png", "static/*.ico"]
"modular.video" = ["PLUGIN.md"]
"modular.voice" = ["PLUGIN.md"]
"modular.trading" = ["PLUGIN.md", "skills/*.md", "agent_templates/*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "e2e_*.py"]
asyncio_default_fixture_loop_scope = "function"
